+++ resources/views/components/public/structured-data.blade.php
@@ -1,0 +1,2 @@
+@props(['data'])
+<script type="application/ld+json">{!! json_encode($data, JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_HEX_QUOT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES) !!}</script>